backend/requirements.txt
flask==2.3.3
sentence-transformers==2.2.2
huggingface_hub==0.7.0 
transformers>=4.20.0
openai==1.3.0
numpy==1.26.0
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.1.0
pytesseract==0.3.10
Pillow==10.0.0
pymongo==4.6.0
pytest==7.4.0
pytest-flask==1.2.0
pytest-mock==3.10.0